% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_summary.R
\name{aux_summary}
\alias{aux_summary}
\title{Auxiliary function for computing standard errors}
\usage{
aux_summary(object, formula, type = "otimist", data, ...)
}
\arguments{
\item{object}{An object of mglm4twin class.}

\item{formula}{formula specifying regression model for each dispersion component.}

\item{type}{Type of standard errors: otimist, robust1 or robust2.}

\item{data}{Data set.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Internal function returns standard errors.
}
\description{
This function provides alternatives ways to compute standard errors.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\keyword{internal}
