% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_S3_methods.R
\name{coef.mglm4twin}
\alias{coef.mglm4twin}
\title{Model Coefficients}
\usage{
\method{coef}{mglm4twin}(object, std.error = FALSE, model, response = NULL, ...)
}
\arguments{
\item{object}{an object of \code{mglm4twin} class.}

\item{std.error}{logical. If \code{TRUE} returns the standard errors
for the estimates. Default is \code{FALSE}.}

\item{model}{Type of the fitted model. Options are \code{E}, \code{AE},
\code{CE}, \code{ACE} and \code{ADE}.}

\item{response}{For which response regression coefficients are extracted.
Default = NULL. It means all estimates are returned including
dispersion estimates for all responses.}

\item{...}{additional arguments affecting the summary produced. Note
that there is no extra options for \code{mglm4twin} object class.}
}
\value{
A \code{data.frame} with parameters names, estimates,
    response variable number and parameters type.
}
\description{
Extract model coefficients for objects
             of \code{mglm4twin} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
