% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_core_pearson.R
\name{ef_core_pearson}
\alias{ef_core_pearson}
\title{Core of the Pearson estimating function.}
\usage{
ef_core_pearson(product, inv_C, res, W)
}
\arguments{
\item{product}{A matrix.}

\item{inv_C}{A matrix.}

\item{res}{A vector of weighted residuals.}

\item{W}{Matrix of weights.}

\item{C}{A matrix.}
}
\value{
A vector.
}
\description{
Core of the Pearson estimating function.
}
\details{
It is an internal function.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
