% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_quasi_score.R
\name{ef_quasi_score}
\alias{ef_quasi_score}
\title{Quasi-score function}
\usage{
ef_quasi_score(D, inv_C, y_vec, mu_vec, W)
}
\arguments{
\item{D}{A matrix. In general the output from
\code{\link[mglm4twin]{mt_link_function}}.}

\item{inv_C}{A matrix. In general the output from
\code{\link[mglm4twin]{mt_build_sigma}}.}

\item{y_vec}{A vector.}

\item{mu_vec}{A vector.}

\item{W}{Matrix of weights.}
}
\value{
The quasi-score vector, the Sensivity and variability
    matrices.
}
\description{
Compute the quasi-score function, its sensitivy and
    variability matrix.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
