% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_sensitivity.R
\name{ef_sensitivity}
\alias{ef_sensitivity}
\title{Sensitivity matrix}
\usage{
ef_sensitivity(product)
}
\arguments{
\item{product}{A list of matrix.}
}
\value{
The sensitivity matrix associated with the Pearson estimating
    function.
}
\description{
Compute the sensitivity matrix associated with the
    Pearson estimating function.
}
\details{
This function implements the equation 7 of Bonat and
    Jorgensen (2016).
}
\author{
Wagner Hugo Bonat and Eduardo Elias Ribeiro Jr
}
\keyword{internal}
