% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anova_mglm.R
\name{mt_anova_mglm}
\alias{mt_anova_mglm}
\title{Wald test for dispersion components}
\usage{
mt_anova_mglm(object, formula, data, ...)
}
\arguments{
\item{object}{An object of mglm4twin class.}

\item{formula}{formula specifying regression model for each dispersion component.}

\item{data}{Data set.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Internal function returns Wald statistics.
}
\description{
This function provides Wald test for dispersion components.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\keyword{internal}
