% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_build_omega.R
\name{mt_build_omega}
\alias{mt_build_omega}
\title{Build omega matrix}
\usage{
mt_build_omega(tau, Z)
}
\arguments{
\item{tau}{A vector}

\item{Z}{A list of matrices in general obtained from mt_twin function.}
}
\value{
A list with the \eqn{\Omega} matrix and its derivatives 
with respect to \eqn{\tau}.
}
\description{
This function builds \eqn{\Omega} matrix.
}
\author{
Wagner Hugo Bonat
}
\keyword{internal}
