% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_compute_h.R
\name{mt_compute_gen}
\alias{mt_compute_gen}
\title{Compute genetic measures and their standard errors}
\usage{
mt_compute_gen(Estimates, vcov, model, n_resp)
}
\arguments{
\item{Estimates}{Table of estimates, standard errors and parameter names.}

\item{vcov}{Matrix of variance and covariance.}

\item{model}{String.}

\item{n_resp}{Numeric. Number of response variables.}
}
\description{
Compute genetic measures and their standard errors
using the delta method.
}
\details{
It is an internal function useful in general for summary
function associated with Twin models.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\keyword{internal}
