% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_compute_rho.R
\name{mt_compute_rho}
\alias{mt_compute_rho}
\title{Compute correlation and its standard errors}
\usage{
mt_compute_rho(Estimates, vcov, component, n_resp)
}
\arguments{
\item{Estimates}{Table of estimates, standard errors and parameter names.}

\item{vcov}{Matrix of variance and covariance.}

\item{component}{Logical indicating for which component the correlation
will be computed.}

\item{n_resp}{Numeric. Number of response variables.}
}
\description{
Compute correlations and standard errors using the delta method.
}
\details{
It is an internal function useful in general for summary
function associated with Twin models.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\keyword{internal}
