% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_delta_method.R
\name{mt_delta_method}
\alias{mt_delta_method}
\title{Delta Method}
\usage{
mt_delta_method(fx, point, cov, std = TRUE)
}
\arguments{
\item{fx}{A function (string like formula) of model parameters.}

\item{point}{Vector of parameter estimates.}

\item{cov}{Variance-covariance matrix.}

\item{std}{Logical. If TRUE returns the standard error, otherwise
return the new variance-covariance.}
}
\description{
Compute standard errors for functions of model parameters
using the delta method.
}
\details{
It is an internal function useful in general for summary
function associated with Twin models.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\keyword{internal}
