% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_derivative_V_sqrt_beta.R
\name{mt_derivative_V_sqrt_beta}
\alias{mt_derivative_V_sqrt_beta}
\title{Derivatives of V^{1/2} with respect to beta.}
\usage{
mt_derivative_V_sqrt_beta(D, D_V_sqrt_mu)
}
\arguments{
\item{D}{A matrix.}

\item{D_V_sqrt_mu}{A matrix.}
}
\value{
A list of matrices, containg the derivatives of \eqn{V^{1/2}}
    with respect to the regression parameters.
}
\description{
Compute the derivatives of \eqn{V^{1/2}} matrix with
    respect to the regression parameters beta.
}
\author{
Wagner Hugo Bonat,  \email{wbonat@ufpr.br}
}
