% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_emp_variability.R
\name{mt_emp_variability}
\alias{mt_emp_variability}
\title{Empirical variability matrix}
\source{
Bonat, W. H. (2018). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, 84(4):1--30.
}
\usage{
mt_emp_variability(object, id)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mglm4twin} class.}

\item{id}{Index indicating independent unit samples.}
}
\value{
Returns a matrix.
}
\description{
Compute an empirical version of the variability matrix
used to estimate robust standard errors.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
