% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_grm.R
\name{mt_grm}
\alias{mt_grm}
\title{Build genetic relationship matrix linear predictor components}
\usage{
mt_grm(n, grm, n_resp, model, formula = NULL, data = NULL)
}
\arguments{
\item{n}{Number of observations.}

\item{grm}{Genetic relationship matrix.}

\item{n_resp}{Number of response variables.}

\item{model}{a string specifying the name of the grm model.
Options are: \code{"E"}, \code{"AE"}}

\item{formula}{formula specifying regression model for each dispersion component.}

\item{data}{Data set.}
}
\value{
A list of matrices.
}
\description{
This function provides the components of the matrix
linear predictor suitable for fitting multivariate genome-based models
by using a precomputed genetic relationship matrix.
}
\author{
Gordon Grabert, \email{wbonat@ufpr.br}
}
\keyword{internal}
