% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_struc.R
\name{mt_struc}
\alias{mt_struc}
\title{Build structural matrix}
\usage{
mt_struc(n_resp)
}
\arguments{
\item{n_resp}{Number of response variables.}
}
\value{
A list of matrices.
}
\description{
This function builds structural matrices to assist
multivariate analysis of twin data. It is an internal function used
in the mc_twin function.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\keyword{internal}
