% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_struc_IPM.R
\name{mt_struc_ipm}
\alias{mt_struc_ipm}
\title{Build structural matrix for independent pathway model}
\usage{
mt_struc_ipm(n_resp)
}
\arguments{
\item{n_resp}{Number of response variables.}
}
\value{
A list of matrices.
}
\description{
This function builds structural matrices to assist
multivariate analysis of twin data. It is an internal function used
in the mc_twin function.
}
\author{
Gordon Grabert, \email{wbonat@ufpr.br}
}
\keyword{internal}
