% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_twin.R
\name{mt_twin}
\alias{mt_twin}
\title{Build twin matrix linear predictor components}
\usage{
mt_twin(N_DZ, N_MZ, n_resp, model, formula = NULL, data = NULL)
}
\arguments{
\item{N_DZ}{Number of DZ twin pairs.}

\item{N_MZ}{Number of MZ twin pairs.}

\item{n_resp}{Number of response variables.}

\item{model}{a string specifying the name of the twin model.
Options are: \code{"E"}, \code{"AE"}, \code{"CE"},
    \code{"ACE"} and \code{"ADE"} models.}

\item{formula}{formula specifying regression model for each dispersion component.}

\item{data}{Data set.}
}
\value{
A list of matrices.
}
\description{
This function provides the components of the matrix
linear predictor suitable for fitting multivariate twin models.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
\keyword{internal}
