% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pAIC.R
\name{pAIC}
\alias{pAIC}
\title{Pseudo Akaike Information Criterion}
\source{
Bonat, W. H. (2018). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, 84(4):1--30.
}
\usage{
pAIC(object, verbose = TRUE)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mglm4twin} class.}

\item{verbose}{logical. Print or not the pAIC value.}
}
\value{
Returns the value of the pseudo Akaike information criterion (pAIC).
}
\description{
Extract the pseudo Akaike information criterion (pAIC)
for objects of \code{mglm4twin} class.
}
\seealso{
\code{gof}, \code{plogLik}, \code{pKLIC} and \code{pBIC}.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
