% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pKLIC.R
\name{pKLIC}
\alias{pKLIC}
\title{Pseudo Kullback-Leibler Information Criterion}
\source{
Bonat, W. H. (2018). Multiple Response Variables Regression
Models in R: The mcglm Package. Journal of Statistical Software, 84(4):1--30.
}
\usage{
pKLIC(object, verbose = TRUE)
}
\arguments{
\item{object}{an object or a list of objects representing a model
of \code{mglm4twin} class.}

\item{verbose}{logical. Print or not the pKLIC value.}
}
\value{
Returns the value of the pseudo Kullback-Leibler information
criterion.
}
\description{
Extract the pseudo Kullback-Leibler information criterion
(pKLIC) for objects of \code{mglm4twin} class.
}
\seealso{
\code{gof}, \code{plogLik}, \code{pAIC}, \code{pBIC} and \code{pKLIC}.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
