% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_S3_methods.R
\name{print.mglm4twin}
\alias{print.mglm4twin}
\title{Print}
\usage{
\method{print}{mglm4twin}(x, ...)
}
\arguments{
\item{x}{fitted model objects of class \code{mglm4twin} as produced
by \code{mglm4twin()}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The default print method for an object of \code{mglm4twin} class.
}
\seealso{
\code{summary}.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
