% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_S3_methods.R
\name{summary.mglm4twin}
\alias{summary.mglm4twin}
\title{Summarizing}
\usage{
\method{summary}{mglm4twin}(object, model, biometric = FALSE, ...)
}
\arguments{
\item{object}{an object of \code{mglm4twin} class.}

\item{model}{String indicating twin model.}

\item{biometric}{Logical. If TRUE biometric measures such as heritability,
common environment, genetic correlation etc are printed. (Default = FALSE).}

\item{...}{additional arguments affecting the summary produced. Note
the there is no extra options for mglm4twin object class.}
}
\value{
Print a \code{mglm4twin} object.
}
\description{
The default summary method for an object of \code{mglm4twin} class.
}
\seealso{
\code{print}.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
