% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_S3_methods.R
\name{vcov.mglm4twin}
\alias{vcov.mglm4twin}
\title{Variance-Covariance Matrix}
\usage{
\method{vcov}{mglm4twin}(object, model, ...)
}
\arguments{
\item{object}{an object of \code{mglm4twin} class.}

\item{model}{String indicating twin model.}

\item{...}{additional arguments affecting the summary produced. Note
that there is no extra options for mcglm object class.}
}
\value{
A variance-covariance matrix.
}
\description{
Returns the variance-covariance matrix for an object
of \code{mglm4twin} class.
}
\author{
Wagner Hugo Bonat, \email{wbonat@ufpr.br}
}
